Rename Product to Item
Excludes the name of fields and endpoints, as this would require bumping
API version. This will come later.
