+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/GetItem.kt
@@ -4,5 +4,8 @@ import no.nb.mlt.wls.domain.model.HostName
 import no.nb.mlt.wls.domain.model.Item
 
 interface GetItem {
-    suspend fun getItem(hostName: HostName, hostId: String): Item?
+    suspend fun getItem(
+        hostName: HostName,
+        hostId: String
+    ): Item?
 }
