+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/synq/SynqError.kt
@@ -3,7 +3,7 @@ package no.nb.mlt.wls.infrastructure.synq
 import org.springframework.web.server.ServerErrorException
 
 data class SynqError(val errorCode: Int, val errorText: String) {
-    class DuplicateProductException(override val cause: Throwable) : ServerErrorException("Product already exists in SynQ", cause)
+    class DuplicateItemException(override val cause: Throwable) : ServerErrorException("Product already exists in SynQ", cause)
 
     class DuplicateOrderException(override val cause: Throwable) : ServerErrorException("Order already exists in SynQ", cause)
 }
